pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
dependencies = [
  'pyaml',
  'setuptools',
  'jinja2',
  'typeguard',
  'numpy',
  'matplotlib',
  'roslibpy',
  'pygame',
  'scipy',
  'colorama',
  'pandas'
]
name = "smartbot-irl"
version = "0.0.1"
authors = [{ name = "Nathaniel pearson", email = "nate@nathanielp.com" }]
description = "A small python wrapper over ROS2 interfaces for smartbot"
readme = "README.md"
requires-python = ">=3.12"
classifiers = [
  "Programming Language :: Python :: 3",
  "Operating System :: OS Independent",
]
license = "MIT"
license-files = ["LICEN[CS]E*"]

[tool.setuptools]
package-dir = { "" = "." }

[tool.setuptools.packages.find]
where = ["."]

[project.urls]
Homepage = "https://github.com/pypa/sampleproject"
Issues = "https://github.com/pypa/sampleproject/issues"
